bug fix
-Env_Item basarili bir sekilde koyuluyor fakat daha eklemem/duzeltmem gereken kucuk seyler var.

+++ src/Others/Utils.cpp
@@ -13,25 +13,23 @@ Utils::Utils()
 
 Utils::~Utils() = default;
 
-
 Vector2<float> Utils::ScreenToWorld(Vector2<float> view, Vector2<float> point)
 {
     return (view + point);
 }
 
-
 Vector2<int> Utils::PositionToCell(Vector2<float> pos)
 {
     if (pos.x >= 0 && pos.y >= 0)
     {
-        return Vector2<int>((int)pos.x / Game_Parameters::SIZE_TILE, (int)pos.y / Game_Parameters::SIZE_TILE);
+        return Vector2<int>(static_cast<int>(pos.x / Game_Parameters::SIZE_TILE), static_cast<int>(pos.y / Game_Parameters::SIZE_TILE));
     }
     return Vector2<int>(-1, -1); //gereksiz
 }
 
 Vector2<float> Utils::CellToPosition(Vector2<int> cell)
 {
-    return Vector2<float>((int)cell.x * Game_Parameters::SIZE_TILE, (int)cell.y * Game_Parameters::SIZE_TILE);
+    return Vector2<float>((static_cast<float>(cell.x)) * Game_Parameters::SIZE_TILE, (static_cast<float>(cell.y)) * Game_Parameters::SIZE_TILE);
 }
 
 int Utils::GenerateID()
